requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
numpy<2.0
pandas
einops
regex

scipy
six

PyYAML
omegaconf
h5py
orjson

colorlog
colorful
tqdm
ipython
ipykernel
line_profiler

pyod
more_itertools
yachalk
ftfy

tensorboardX
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1
transformers

Pillow
opencv-python
ffmpeg-python
decord

spacy
stanza
spacy_stanza
